Light refactoring

+++ src/Utils/Classes/Collections/ClassedCollection.php
@@ -31,26 +31,19 @@ class ClassedCollection implements Countable, Iterator {
 	}
 
 	public function array(): array {
-		$array = array();
-
-		foreach($this as $object) {
-			array_push($array, $object);
-		}
-
-		return $array;
+		return iterator_to_array($this, preserve_keys: false);
 	}
 
 	public function get(string $class): ClassedCollection {
-		$objects = new ClassedCollection($this->class);
-
 		// Search all objects for one matching the given class or subclass
-		foreach($this as $object) {
-			if(is_a($object, $class)) {
-				$objects->add($object);
-			}
-		}
+		$objects = array_filter($this->array(), fn($object) => is_a($object, $class));
 
-		return $objects;
+		// Convert back to ClassedCollection
+		$collection = new ClassedCollection($this->class);
+		$collection->add(...$objects);
+
+		// Return the result
+		return $collection;
 	}
 
 	public function first(string $class): object {
@@ -72,12 +65,11 @@ class ClassedCollection implements Countable, Iterator {
 	public function add(object ...$objects): void {
 		foreach($objects as $object) {
 			// Assert that the given object matches the type of this ClassedCollection
-			$expected_class = $this->class;
-			$object_class = $object::class;
-			assert(
-				is_a($object, $this->class),
-				new InvalidArgumentException("Failed while adding object to collection: Cannot add object of class '$object_class' to collection of '$expected_class'")
-			);
+			if(!is_a($object, $this->class)) {
+				$expected_class = $this->class;
+				$object_class = $object::class;
+				throw new InvalidArgumentException("Failed while adding object to collection: Cannot add object of class '$object_class' to collection of '$expected_class'");
+			}
 
 			$this->data->attach($object);
 		}
@@ -88,20 +80,16 @@ class ClassedCollection implements Countable, Iterator {
 			$object = $object_or_class;
 
 			// Assert that the given object matches the type of this ClassedCollection
-			$expected_class = $this->class;
-			$object_class = $object::class;
-			assert(
-				is_a($object, $this->class),
-				new InvalidArgumentException("Failed while removing object from collection: Object of class '$object_class' cannot be in collection of '$expected_class'")
-			);
+			if(!is_a($object, $this->class)) {
+				$expected_class = $this->class;
+				$object_class = $object::class;
+				throw new InvalidArgumentException("Failed while removing object from collection: Object of class '$object_class' cannot be in collection of '$expected_class'");
+			}
 
 			$this->data->detach($object);
 		} else {
 			$class = $object_or_class;
-
-			foreach($this->get($class) as $object) {
-				$this->remove($object);
-			}
+			array_map(fn($object) => $this->remove($object), $this->get($class)->array());
 		}
 	}
